+++ src/main/java/dev/lambdaurora/lambdynlights/mixin/lightsource/LivingEntityMixin.java
@@ -40,7 +40,7 @@ public abstract class LivingEntityMixin extends Entity implements DynamicLightSo
 			this.lambdynlights$luminance = 15;
 		} else {
 			int luminance = 0;
-			var eyePos = new BlockPos(this.getX(), this.getEyeY(), this.getZ());
+			var eyePos = new BlockPos((int) this.getX(), (int) this.getEyeY(), (int) this.getZ());
 			boolean submergedInFluid = !this.world.getFluidState(eyePos).isEmpty();
 			for (var equipped : this.getItemsEquipped()) {
 				if (!equipped.isEmpty())
